feat: initial rudimentary implementation of logging
Should provide a basic and stable set of primitives for logging.
