+++ react/index.ts
@@ -1,5 +1,5 @@
 export * from "./components";
 export * from "./app";
-export * from "./log";
+export * from "./logging";
 export * from "./structs";
 export * from "./util";
